Move traits to dedicated module
With the introduction of `Driver`, they are no longer as relevant to end
users as they were. I think it makes sense to move them to their own
module, instead of cluttering the top-level namespace.
