+++ src/lib.rs
@@ -28,55 +28,16 @@ pub extern crate embedded_time;
 
 /// Re-exports the traits from this library
 pub mod prelude {
-    pub use super::{SetStepMode as _, Step as _};
+    pub use crate::traits::{SetStepMode as _, Step as _};
 }
 
 pub mod drivers;
+pub mod traits;
 
 mod driver;
 mod step_mode;
 
 pub use self::{driver::*, step_mode::*};
-
-use embedded_time::Clock;
-
-/// Blocking interface for setting the step mode
-pub trait SetStepMode {
-    /// The error that can occur while using this trait
-    type Error;
-
-    /// The type that defines the microstepping mode
-    ///
-    /// This crate includes a number of enums that can be used for this purpose.
-    type StepMode: StepMode;
-
-    /// Sets the step mode
-    fn set_step_mode<Clk: Clock>(
-        &mut self,
-        step_mode: Self::StepMode,
-        clock: &Clk,
-    ) -> Result<(), Self::Error>;
-}
-
-/// Blocking interface for making single steps
-pub trait Step {
-    /// The error that can occur while using this trait
-    type Error;
-
-    /// Rotates the motor one (micro-)step in the given direction
-    ///
-    /// This should result in the motor making one step. To achieve a specific
-    /// speed, the user must call this method at the appropriate frequency.
-    ///
-    /// Requires a reference to an `embedded_time::Clock` implementation to
-    /// handle the timing. Please make sure that the timer doesn't overflow
-    /// while this method is running.
-    fn step<Clk: Clock>(
-        &mut self,
-        dir: Dir,
-        clock: &Clk,
-    ) -> Result<(), Self::Error>;
-}
 
 /// Defines the direction in which to rotate the motor
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
